+++ backend/src/test/java/com/allog/dallog/categoryrole/application/CategoryRoleServiceTest.java
@@ -1,5 +1,10 @@
-package com.allog.dallog.domain.categoryrole.application;
+package com.allog.dallog.categoryrole.application;
 
+import static com.allog.dallog.category.domain.CategoryType.GOOGLE;
+import static com.allog.dallog.category.domain.CategoryType.NORMAL;
+import static com.allog.dallog.category.domain.CategoryType.PERSONAL;
+import static com.allog.dallog.categoryrole.domain.CategoryRoleType.ADMIN;
+import static com.allog.dallog.categoryrole.domain.CategoryRoleType.NONE;
 import static com.allog.dallog.common.Constants.개인_카테고리_이름;
 import static com.allog.dallog.common.Constants.나인_이름;
 import static com.allog.dallog.common.Constants.나인_이메일;
@@ -9,31 +14,26 @@ import static com.allog.dallog.common.Constants.취업_카테고리_이름;
 import static com.allog.dallog.common.Constants.티거_이름;
 import static com.allog.dallog.common.Constants.티거_이메일;
 import static com.allog.dallog.common.Constants.티거_프로필_URL;
-import static com.allog.dallog.domain.category.domain.CategoryType.GOOGLE;
-import static com.allog.dallog.domain.category.domain.CategoryType.NORMAL;
-import static com.allog.dallog.domain.category.domain.CategoryType.PERSONAL;
-import static com.allog.dallog.domain.categoryrole.domain.CategoryRoleType.ADMIN;
-import static com.allog.dallog.domain.categoryrole.domain.CategoryRoleType.NONE;
-import static com.allog.dallog.domain.subscription.domain.Color.COLOR_1;
+import static com.allog.dallog.subscription.domain.Color.COLOR_1;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.allog.dallog.category.domain.Category;
+import com.allog.dallog.category.domain.CategoryRepository;
+import com.allog.dallog.category.domain.CategoryType;
+import com.allog.dallog.categoryrole.domain.CategoryRole;
+import com.allog.dallog.categoryrole.domain.CategoryRoleRepository;
+import com.allog.dallog.categoryrole.dto.request.CategoryRoleUpdateRequest;
+import com.allog.dallog.categoryrole.dto.response.SubscribersResponse;
+import com.allog.dallog.categoryrole.exception.ManagingCategoryLimitExcessException;
+import com.allog.dallog.categoryrole.exception.NoCategoryAuthorityException;
+import com.allog.dallog.categoryrole.exception.NotAbleToChangeRoleException;
 import com.allog.dallog.common.annotation.ServiceTest;
-import com.allog.dallog.domain.category.domain.Category;
-import com.allog.dallog.domain.category.domain.CategoryRepository;
-import com.allog.dallog.domain.category.domain.CategoryType;
-import com.allog.dallog.domain.categoryrole.domain.CategoryRole;
-import com.allog.dallog.domain.categoryrole.domain.CategoryRoleRepository;
-import com.allog.dallog.domain.categoryrole.dto.request.CategoryRoleUpdateRequest;
-import com.allog.dallog.domain.categoryrole.dto.response.SubscribersResponse;
-import com.allog.dallog.domain.categoryrole.exception.ManagingCategoryLimitExcessException;
-import com.allog.dallog.domain.categoryrole.exception.NoCategoryAuthorityException;
-import com.allog.dallog.domain.categoryrole.exception.NotAbleToChangeRoleException;
-import com.allog.dallog.domain.member.domain.Member;
-import com.allog.dallog.domain.member.domain.MemberRepository;
-import com.allog.dallog.domain.member.domain.SocialType;
-import com.allog.dallog.domain.subscription.domain.Subscription;
-import com.allog.dallog.domain.subscription.domain.SubscriptionRepository;
+import com.allog.dallog.member.domain.Member;
+import com.allog.dallog.member.domain.MemberRepository;
+import com.allog.dallog.member.domain.SocialType;
+import com.allog.dallog.subscription.domain.Subscription;
+import com.allog.dallog.subscription.domain.SubscriptionRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
