+++ backend/src/test/java/com/allog/dallog/externalcalendar/application/ExternalCalendarServiceTest.java
@@ -1,4 +1,4 @@
-package com.allog.dallog.domain.externalcalendar.application;
+package com.allog.dallog.externalcalendar.application;
 
 import static com.allog.dallog.common.Constants.나인_이름;
 import static com.allog.dallog.common.Constants.나인_이메일;
@@ -6,12 +6,12 @@ import static com.allog.dallog.common.Constants.나인_프로필_URL;
 import static com.allog.dallog.common.fixtures.OAuthTokenFixtures.OAUTH_TOKEN;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.allog.dallog.auth.domain.OAuthTokenRepository;
 import com.allog.dallog.common.annotation.ServiceTest;
-import com.allog.dallog.domain.auth.domain.OAuthTokenRepository;
-import com.allog.dallog.domain.externalcalendar.dto.ExternalCalendarsResponse;
-import com.allog.dallog.domain.member.domain.Member;
-import com.allog.dallog.domain.member.domain.MemberRepository;
-import com.allog.dallog.domain.member.domain.SocialType;
+import com.allog.dallog.externalcalendar.dto.ExternalCalendarsResponse;
+import com.allog.dallog.member.domain.Member;
+import com.allog.dallog.member.domain.MemberRepository;
+import com.allog.dallog.member.domain.SocialType;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
